refactor: move SSO callbacks to routes.ts and consolidate route handling
- Add SSO callback routes to routes.ts configuration
- Update login.tsx and signup.tsx to handle SSO callback paths
- Remove separate SSO callback route files
- Use pathname detection to show processing UI for callbacks
- Maintain single route files with conditional rendering

Better architecture following React Router patterns.
