+++ vibes.diy/pkg/app/routes/login.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from "react";
-import { useNavigate } from "react-router-dom";
+import { useNavigate, useLocation } from "react-router-dom";
 import { ClerkAuthProvider, VibesClerkAuth } from "use-vibes";
 import SimpleAppLayout from "../components/SimpleAppLayout.js";
 import { useAuth } from "../contexts/AuthContext.js";
@@ -17,7 +17,11 @@ export function meta() {
 
 function LoginContent() {
   const navigate = useNavigate();
+  const location = useLocation();
   const { isAuthenticated: isVibesAuth } = useAuth();
+  
+  // Check if this is an SSO callback route
+  const isSSCallback = location.pathname.includes('/sso-callback');
 
   // If user is already authenticated via existing system, redirect home
   useEffect(() => {
@@ -32,6 +36,39 @@ function LoginContent() {
     // Navigate home - user is now fully authenticated via Fireproof integration
     navigate("/");
   };
+
+  // Handle SSO callback case - show processing UI
+  if (isSSCallback) {
+    useEffect(() => {
+      // Give Clerk a moment to process the callback, then redirect to home
+      const timer = setTimeout(() => {
+        navigate("/");
+      }, 1000);
+
+      return () => clearTimeout(timer);
+    }, [navigate]);
+
+    return (
+      <SimpleAppLayout>
+        <div className="flex min-h-screen items-center justify-center py-12 px-4">
+          <div className="w-full max-w-md space-y-8 text-center">
+            <div>
+              <h2 className="mt-6 text-3xl font-bold tracking-tight text-gray-900 dark:text-white">
+                Processing authentication...
+              </h2>
+              <p className="mt-2 text-sm text-gray-600 dark:text-gray-400">
+                Please wait while we complete your sign-in
+              </p>
+            </div>
+
+            <div className="mt-8">
+              <div className="mx-auto h-2 w-24 animate-pulse rounded-full bg-orange-500" />
+            </div>
+          </div>
+        </div>
+      </SimpleAppLayout>
+    );
+  }
 
   return (
     <SimpleAppLayout>
